tidy up the search bar
